db connect, ci/cd
