+++ .github/workflows/ci.yml
@@ -1,0 +1,29 @@
+name: Java CI with Gradle
+
+on:
+  push:
+    branches: [ "main" ]
+  pull_request:
+    branches: [ "main" ]
+
+permissions:
+  contents: read
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v4
+    - name: Set up JDK 21
+      uses: actions/setup-java@v4
+      with:
+        java-version: '21'
+        distribution: 'temurin'
+    - name: Grant execute permission for gradlew
+      run: chmod +x gradlew
+    - name: Build with Gradle
+      uses: gradle/actions/setup-gradle@v3
+      with:
+        build-root-directory: C:\Users\shito\9oorm\2025_SEASONTHON_TEAM_97_BE
+    - name: Build with Gradle
+      run: ./gradlew build
