--- ccsketch/src/main/java/com/example/ccsketch/CcsketchApplication.java
@@ -1,15 +1,0 @@
-package com.example.ccsketch;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@SpringBootApplication
-@EnableJpaAuditing
-public class CcsketchApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CcsketchApplication.class, args);
-	}
-
-}
