--- ccsketch/src/main/java/com/example/ccsketch/DTO/LoginRequestDto.java
@@ -1,13 +1,0 @@
-package com.example.ccsketch.DTO;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class LoginRequestDto {
-    private String loginId;
-    private String password;
-}
-
-
