--- ccsketch/src/main/java/com/example/ccsketch/DTO/LoginResponseDto.java
@@ -1,12 +1,0 @@
-package com.example.ccsketch.DTO;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class LoginResponseDto {
-    private String accessToken;
-}
-
-
