--- ccsketch/src/main/java/com/example/ccsketch/DTO/SearchRequestDto.java
@@ -1,16 +1,0 @@
-package com.example.ccsketch.DTO;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-
-public class SearchRequestDto {
-    @Schema(description = "검색할 키워드", required = true, example = "청년 창업")
-    private String keyword;
-
-    public String getKeyword() {
-        return keyword;
-    }
-
-    public void setKeyword(String keyword) {
-        this.keyword = keyword;
-    }
-}
