--- ccsketch/src/main/java/com/example/ccsketch/DTO/SignUpRequestDto.java
@@ -1,18 +1,0 @@
-package com.example.ccsketch.DTO;
-
-import lombok.Getter;
-import lombok.Setter;
-import java.time.LocalDate;
-
-@Getter
-@Setter
-public class SignUpRequestDto {
-    private String loginId;
-    private String password;
-    private String name;
-    private LocalDate birth;
-    private String location;
-    private String organizationStatus;
-    private Integer financialStatus;
-    private Integer goal;
-}
