--- ccsketch/src/main/java/com/example/ccsketch/util/JwtTokenProvider.java
@@ -1,48 +1,0 @@
-package com.example.ccsketch.util;
-
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-import java.nio.charset.StandardCharsets;
-import javax.crypto.SecretKey;
-import java.util.Date;
-
-@Component
-public class JwtTokenProvider {
-
-    private final SecretKey signingKey;
-    private final long expirationMillis;
-
-    public JwtTokenProvider(
-            @Value("${jwt.secret}") String secret,
-            @Value("${jwt.expiration-ms}") long expirationMillis
-    ) {
-        this.signingKey = Keys.hmacShaKeyFor(secret.getBytes(StandardCharsets.UTF_8));
-        this.expirationMillis = expirationMillis;
-    }
-
-    public String createToken(String subjectEmail) {
-        Date now = new Date();
-        Date expiry = new Date(now.getTime() + expirationMillis);
-
-        return Jwts.builder()
-                .subject(subjectEmail)
-                .issuedAt(now)
-                .expiration(expiry)
-                .signWith(signingKey)
-                .compact();
-    }
-
-    public String getSubject(String token) {
-        return Jwts.parser()
-                .verifyWith(signingKey)
-                .build()
-                .parseSignedClaims(token)
-                .getPayload()
-                .getSubject();
-    }
-}
-
-
