--- ccsketch/src/test/java/com/example/ccsketch/CcsketchApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.ccsketch;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CcsketchApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
