+++ src/main/java/com/example/ccsketch/api/dto/AiSearchResponseDto.java
@@ -1,0 +1,23 @@
+package com.example.ccsketch.api.dto;
+
+import java.util.List;
+
+public class AiSearchResponseDto {
+    private List<Long> resultCardIds;
+    private String modelVersion;
+    public List<Long> getResultCardIds() {
+        return resultCardIds;
+    }
+
+    public void setResultCardIds(List<Long> resultCardIds) {
+        this.resultCardIds = resultCardIds;
+    }
+
+    public String getModelVersion() {
+        return modelVersion;
+    }
+
+    public void setModelVersion(String modelVersion) {
+        this.modelVersion = modelVersion;
+    }
+}
