+++ src/main/java/com/example/ccsketch/domain/user/dto/request/LoginRequestDto.java
@@ -1,0 +1,11 @@
+package com.example.ccsketch.domain.user.dto.request;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class LoginRequestDto {
+    private String loginId;
+    private String password;
+}
