+++ src/main/java/com/example/ccsketch/domain/user/dto/response/LoginResponseDto.java
@@ -1,0 +1,10 @@
+package com.example.ccsketch.domain.user.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class LoginResponseDto {
+    private String accessToken;
+}
