+++ src/main/resources/application.properties
@@ -1,1 +1,41 @@
 spring.application.name=youth_sketch_BE
+
+
+# MariaDB
+# H2
+spring.h2.console.enabled=true
+spring.h2.console.path=/h2-console
+spring.datasource.url=jdbc:h2:mem:testdb
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=
+
+# JPA
+spring.jpa.hibernate.ddl-auto=create
+spring.jpa.show-sql=true
+
+# Kakao OAuth2 Settings
+spring.security.oauth2.client.registration.kakao.client-id=8d5698a9a539093baad3fc290048dc1f
+# client-secret은 비워둬도 괜찮습니다 (필수 아님)
+# spring.security.oauth2.client.registration.kakao.client-secret=
+spring.security.oauth2.client.registration.kakao.redirect-uri={baseUrl}/{action}/oauth2/code/{registrationId}
+spring.security.oauth2.client.registration.kakao.authorization-grant-type=authorization_code
+spring.security.oauth2.client.registration.kakao.client-authentication-method=client_secret_post
+spring.security.oauth2.client.registration.kakao.scope=profile_nickname,account_email
+
+spring.security.oauth2.client.provider.kakao.authorization-uri=https://kauth.kakao.com/oauth/authorize
+spring.security.oauth2.client.provider.kakao.token-uri=https://kauth.kakao.com/oauth/token
+spring.security.oauth2.client.provider.kakao.user-info-uri=https://kapi.kakao.com/v2/user/me
+spring.security.oauth2.client.provider.kakao.user-name-attribute=id
+
+# JWT Settings
+jwt.secret=ThisIsADevelopmentOnlySecretKeyForJwtMustBeAtLeast32Bytes!!
+jwt.expiration-ms=3600000
+
+# Disable Spring AI (OpenAI) in default profile
+spring.ai.openai.enabled=false
+spring.ai.openai.chat.enabled=false
+spring.ai.openai.image.enabled=false
+spring.ai.openai.speech.enabled=false
+spring.ai.openai.embedding.enabled=false
+spring.autoconfigure.exclude=org.springframework.ai.autoconfigure.vectorstore.redis.RedisVectorStoreAutoConfiguration,org.springframework.ai.autoconfigure.vectorstore.pgvector.PgVectorStoreAutoConfiguration,org.springframework.ai.autoconfigure.vectorstore.cassandra.CassandraVectorStoreAutoConfiguration,org.springframework.ai.autoconfigure.vectorstore.mongo.MongoDBAtlasVectorStoreAutoConfiguration,org.springframework.ai.autoconfigure.vectorstore.qdrant.QdrantVectorStoreAutoConfiguration,org.springframework.ai.autoconfigure.vectorstore.chroma.ChromaVectorStoreAutoConfiguration,org.springframework.ai.autoconfigure.vectorstore.milvus.MilvusVectorStoreAutoConfiguration,org.springframework.ai.autoconfigure.vectorstore.mssql.MssqlVectorStoreAutoConfiguration,org.springframework.ai.autoconfigure.vectorstore.oracle.OracleVectorStoreAutoConfiguration,org.springframework.ai.autoconfigure.vectorstore.weaviate.WeaviateVectorStoreAutoConfiguration,org.springframework.ai.autoconfigure.vectorstore.elasticsearch.ElasticsearchVectorStoreAutoConfiguration,org.springframework.ai.autoconfigure.vectorstore.weaviate.WeaviateVectorStoreEasyAutoConfiguration,org.springframework.ai.autoconfigure.retry.RetryAutoConfiguration,org.springframework.ai.autoconfigure.observation.ChatClientObservationAutoConfiguration,org.springframework.ai.autoconfigure.http.HttpAutoConfiguration,org.springframework.ai.autoconfigure.vectorstore.FailedVectorStoreAutoConfiguration,org.springframework.ai.model.openai.autoconfigure.OpenAiAudioSpeechAutoConfiguration,org.springframework.ai.model.openai.autoconfigure.OpenAiImageAutoConfiguration,org.springframework.ai.model.openai.autoconfigure.OpenAiChatModelAutoConfiguration,org.springframework.ai.model.openai.autoconfigure.OpenAiChatAutoConfiguration,org.springframework.ai.model.openai.autoconfigure.OpenAiEmbeddingAutoConfiguration,org.springframework.ai.model.openai.autoconfigure.OpenAiModerationAutoConfiguration,org.springframework.ai.autoconfigure.api.ApiAutoConfiguration,org.springframework.ai.model.openai.autoconfigure.OpenAiAudioTranscriptionAutoConfiguration
