+++ src/test/resources/application-test.properties
@@ -1,0 +1,12 @@
+# JWT Settings for Test
+jwt.secret=ThisIsADevelopmentOnlySecretKeyForJwtMustBeAtLeast32BytesLongForTest
+jwt.expiration-ms=3600000
+
+# H2 Test DB
+spring.datasource.url=jdbc:h2:mem:testdb
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=
+
+# Disable OpenAI for tests by providing a dummy key
+spring.ai.openai.api-key=dummy-key
